Made ComInfo.id() more robust.
The ComID is created on-the-fly when ComInfo.id() is called,
which prevents it from being invalid.

Obviously, this means it must perform a lookup in the component list,
which is a binary search over a small and hot vector.

If the lookup becomes a problem, the solution may be to allow visitors
to accept a ComID directly, which would match the component of the
previous visitor parameter (e.g. [](ComInfo<Sprite> info, ComID id){}).